requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.4
deepgram-sdk==3.1.3
Django==5.0.2
djangorestframework==3.14.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
marshmallow==3.20.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
packaging==23.2
pytz==2024.1
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
sympy==1.12
tiktoken==0.5.2
torch==2.2.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
verboselogs==1.7
websockets==12.0
word2number==1.1
yarl==1.9.4
